add hdfs datanode count to operational stats
