+++ cdap-operational-stats-core/pom.xml
@@ -28,6 +28,10 @@
   <artifactId>cdap-operational-stats-core</artifactId>
   <name>CDAP Operational Stats Core</name>
 
+  <properties>
+    <httpclient.version>4.5.2</httpclient.version>
+  </properties>
+
   <dependencies>
     <dependency>
       <groupId>co.cask.cdap</groupId>
@@ -119,6 +123,14 @@
       <groupId>org.slf4j</groupId>
       <artifactId>jcl-over-slf4j</artifactId>
       <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.httpcomponents</groupId>
+      <artifactId>httpclient</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.httpcomponents</groupId>
+      <artifactId>httpcore</artifactId>
     </dependency>
   </dependencies>
